Directed graph creation(describes input file needed)
Input is based on position, it is values space out by white space.
If N is the number of nodes the input must have N*N values. The first N
values are the flow values for Node 1 to any other given node. If there
is no connection put in the value of 0.